Change epoch / position of leafs to u32
As we only consider Merkle trees with up to 2^32
leafs, we can assume epochs and positions in layers
in trees fit into a u32.

+++ benches/benchmark.rs
@@ -64,7 +64,7 @@ pub fn benchmark_signature_scheme<S: SignatureScheme>(c: &mut Criterion, descrip
             rng.fill(&mut message);
 
             // Sample random epoch
-            let epoch = rng.gen_range(0..S::LIFETIME) as u64;
+            let epoch = rng.gen_range(0..S::LIFETIME) as u32;
 
             // Benchmark signing
             let _ = S::sign(
@@ -77,11 +77,11 @@ pub fn benchmark_signature_scheme<S: SignatureScheme>(c: &mut Criterion, descrip
     });
 
     // Pre-generate messages, epochs, and signatures for verification
-    let precomputed: Vec<(u64, [u8; MESSAGE_LENGTH], S::Signature)> = (0..2000)
+    let precomputed: Vec<(u32, [u8; MESSAGE_LENGTH], S::Signature)> = (0..2000)
         .map(|_| {
             let mut message = [0u8; MESSAGE_LENGTH];
             rng.fill(&mut message);
-            let epoch = rng.gen_range(0..S::LIFETIME) as u64;
+            let epoch = rng.gen_range(0..S::LIFETIME) as u32;
             let signature =
                 S::sign(&mut rng, &sk, epoch, &message).expect("Signing should succeed");
             (epoch, message, signature)
